pyproject.toml
[tool.poetry]
name = "homework_bot"
version = "0.1.0"
description = ""
authors = ["SaPer663 <saper663ru@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "0.19.0"
python-telegram-bot = "13.7"
requests = "2.26.0"

[tool.poetry.dev-dependencies]
flake8 = "3.9.2"
flake8-docstrings = "1.6.0"
pytest = "6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
